完善数据库表结构，添加房屋状态，用以判断是否又业主入住，员工表增加入职时间

+++ test_project1/src/main/java/com/guodong/core/dao/ManagementfeeDao.java
@@ -11,7 +11,7 @@ public interface ManagementfeeDao {
 
     int deleteByExample(ManagementfeeQuery example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Managementfee record);
 
@@ -19,7 +19,7 @@ public interface ManagementfeeDao {
 
     List<Managementfee> selectByExample(ManagementfeeQuery example);
 
-    Managementfee selectByPrimaryKey(Integer id);
+    Managementfee selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") Managementfee record, @Param("example") ManagementfeeQuery example);
